SCV/requirement.txt
streamlit==1.28.0
pdfplumber==0.11.8
python-docx==0.8.11
scikit-learn==1.3.2
nltk==3.8.1
google-generativeai==0.8.3
pandas==2.2.2
tqdm==4.66.0
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
python-dotenv==1.0.0
